requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
requests==2.30.0
six==1.16.0
sniffio==1.3.0
tzlocal==5.0.1
urllib3==2.0.2
yarl==1.9.2
